# swagger.yaml
openapi: 3.0.0
paths:
...
  /signin:
    post:
      operationId: AuthController_signin
      summary: Войти в систему
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: test@mail.ru
                password:
                  type: string
                  example: '123'
      responses:
        '201':
          description: Успешный вход в систему
          content:
            application/json:
              schema:
                type: object
                properties:
                  _id:
                    type: string
                    description: Идентификатор
                    example: 650b396dd4201e5ca499f3b3
                  username:
                    type: string
                    description: Имя пользователя
                    example: test
                  phone:
                    type: string
                    description: Номер телефона
                    example: '+79999999999'
                  avatar:
                    type: string
                    description: URL аватара
                    example: https://i.pravatar.cc/300
                  balance:
                    type: number
                    description: Баланс
                    example: 0
                  accounts:
                    type: array
                    items:
                      type: object
                      properties:
                        _id:
                          type: string
                          description: Идентификатор аккаунта
                          example: 650b396ed4201e5ca499f3b5
                        type:
                          type: string
                          description: Тип аккаунта
                          example: local
                        role:
                          type: string
                          description: Роль аккаунта
                          example: user
                        credentials:
                          type: object
                          properties:
                            email:
                              type: string
                              description: Email
                              example: test@mail.ru
                            accessToken:
                              type: string
                              description: AccessToken
                              example: >-
                                eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiI2NTBiMzk2ZGQ0MjAxZTVjYTQ5OWYzYjMiLCJpYXQiOjE2OTUyMzQ0MTQsImV4cCI6MTY5NTMyMDgxNH0.1GIu8iWeg8iWF-i5iynAhelc7kO3ouj09boZHjqn5HE
                            refreshToken:
                              type: string
                              description: RefreshToken
                              example: eyJhbGciOiJIUzI1NiIsInR5cCI6Ikp
                        profile:
                          type: string
                          description: Идентификатор профиля
                          example: 650b396dd4201e5ca499f3b3
                    description: Список аккаунтов
        '401':
          description: Неверное имя пользователя или пароль
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Неверное имя пользователя или пароль
                    description: Сообщение об ошибке
                  error:
                    type: string
                    example: Unauthorized
                    description: Тип ошибки
                  statusCode:
                    type: number
                    example: '401'
                    description: HTTP-статус код
      tags: &ref_0
        - auth
  /signup:
    post:
      operationId: AuthController_signup
      summary: Регистрация
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                password:
                  type: string
                  example: '123'
                email:
                  type: string
                  example: test@mail.ru
                phone:
                  type: string
                  example: '+79999999999'
                username:
                  type: string
                  example: test
      responses:
        '201':
          description: Успешная регистрация
          content:
            application/json:
              schema:
                type: object
                properties:
                  username:
                    type: string
                    description: Имя пользователя
                    example: test
                  phone:
                    type: string
                    description: Номер телефона
                    example: '+79999999999'
                  avatar:
                    type: string
                    description: URL аватара
                    example: https://i.pravatar.cc/300
                  balance:
                    type: number
                    description: Баланс
                    example: 0
                  accounts:
                    type: array
                    items:
                      type: object
                      properties:
                        type:
                          type: string
                          description: Тип аккаунта
                          example: local
                        role:
                          type: string
                          description: Роль
                          example: user
                        credentials:
                          type: object
                          properties:
                            email:
                              type: string
                              description: Email
                              example: test@mail.ru
                            accessToken:
                              type: string
                              description: AccessToken
                              example: eyJhbGciOiJIUzI1NiIsInR5cCI6Ikp
                            refreshToken:
                              type: string
                              description: RefreshToken
                              example: eyJhbGciOiJIUzI1NiIsInR5cCI6Ikp
                        profile:
                          type: string
                          description: 'Идентификатор профиля '
                          example: 650b396dd4201e5ca499f3b3
                        _id:
                          type: string
                          description: Идентификатор аккаунта
                          example: 650b396dd4201e5ca499f3b3
                    description: Список аккаунтов
                  _id:
                    type: string
                    description: Идентификатор профиля
                    example: 650b396dd4201e5ca499f3b3
        '400':
          description: Некорректные данные
        '409':
          description: Аккаунт уже существует
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Аккаунт уже существует
                    description: Сообщение об ошибке
                  error:
                    type: string
                    example: Conflict
                    description: Тип ошибки
                  statusCode:
                    type: number
                    example: '409'
                    description: HTTP-статус код
      tags: *ref_0
  /refresh-token:
    post:
      operationId: AuthController_refreshToken
      summary: Обновить токен
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  type: string
                  example: eyJhbGciOiJIUzI1NiIsInR5cCI6Ikp
      responses:
        '201':
          description: accessToken успешно обновлен
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
                    description: accessToken по умолчанию действует 1 день
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6Ikp
                  refreshToken:
                    type: string
                    description: refreshToken по умолчанию действует 7 дней
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6Ikp
        '401':
          description: Невалидный refreshToken
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Невалидный refreshToken
                    description: Сообщение об ошибке
                  error:
                    type: string
                    example: Unauthorized
                    description: Тип ошибки
                  statusCode:
                    type: number
                    example: '401'
                    description: HTTP-статус код
      tags: *ref_0
...
  /yandex/success:
    get:
      operationId: AuthController_yandexSuccess
      summary: ''
      parameters:
        - name: token
          required: true
          in: query
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_0
...
  /reset-password:
    post:
      operationId: AuthController_resetPassword
      summary: Сброс пароля
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  description: Email пользователя
                  example: test@mail.ru
      responses:
        '201':
          description: Ссылка для сброса пароля отправлена на указанный Email
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: >-
                      Ссылка на сброс пароля отправлена на ваш email:
                      test@mail.ru
                    description: Email пользователя
        '400':
          description: Некорректные данные
        '404':
          description: Пользователь с указанным Email не найден
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Пользователь с указанным Email не найден
                    description: Сообщение об ошибке
                  error:
                    type: string
                    example: Not Found
                    description: Тип ошибки
                  statusCode:
                    type: number
                    example: '404'
                    description: HTTP-статус код
      tags: *ref_0
...
info:
  title: API BotKits
  description: Ручки для команды frontend
  version: '1.0'
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
...